add verify

+++ revert.rb
@@ -36,6 +36,11 @@ Dir.glob(File.join(in_dir, '*.jpg')) do |f|
     new_path = File.join(new_path, $2)
     puts "move to #{new_path}"
     FileUtils.cp(f, new_path)
+    unless Digest::MD5.file(new_path).hexdigest == Digest::MD5.file(f).hexdigest
+      puts Digest::MD5.hexdigest(new_path)
+      puts Digest::MD5.hexdigest(f)
+      raise "#{f} move verify error"
+    end
     FileUtils.rm(f)
   else
     puts "sikip #{f}"
